ChemSim/requirements-dev.txt
pygame>=2.5.0
PyYAML>=6.0.1
pytest>=8.0.0
